discovered error in prior renaming.  BehaviorInput and BehaviorInputContainer now  have correct naming conventions.
