+++ client/src/Behavior.jsx
@@ -1,13 +1,18 @@
 import React, { useState, useEffect } from 'react';
-import BehaviorInput from './components/BehaviorInputContainer';
-import BehaviorContainer from './components/BehaviorContainer';
+import BehaviorInput from './components/BehaviorInput';
+import BehaviorInputContainer from './components/BehaviorContainer';
 
-const emptyPhase = { rounds: 0, orders: 0 };
+// empty phase data for reset ease
+const emptyPhase = {
+    rounds: 0, // represents numbers of rounds to place the number of orders
+    orders: 0 // repreents number of orders to place per round
+};
 
+//Behavior view component
 const Behavior = ({ rounds, onSubmit, onCancel, isClicked }) => {
     //is this needed?
     const [maxRounds, setMaxRounds] = useState(rounds ? typeof rounds === 'number' ? rounds : 10 : 10)
-    //phase objs, max 3 phases of behavior
+    //phase objs, max 3 phases of behavior, all default to 0 values
     const [phase1, setPhase1] = useState({ rounds: 0, orders: 0 });
     const [phase2, setPhase2] = useState({ rounds: 0, orders: 0 });
     const [phase3, setPhase3] = useState({ rounds: 0, orders: 0 });
@@ -15,18 +20,23 @@ const Behavior = ({ rounds, onSubmit, onCancel, isClicked }) => {
     //this was the beginning of exploring hiding behavior input until custom behavior element is clicked
     const [isPhase1Disabled, setIsPhase1Disabled] = useState(isClicked ? isClicked : true);
 
+    //controls phase2/3 visibility, defaults to true
     const [isPhase2Disabled, setIsPhase2Disabled] = useState(true);
     const [isPhase3Disabled, setIsPhase3Disabled] = useState(true);
 
+    //gets max rounds for phase2
     const phase2MaxRounds = () => {
         const max = (maxRounds - phase1.rounds) ? phase1.rounds : 0;
         return max;
     }
+
+    //gets max rounds for phase3
     const phase3MaxRounds = () => {
         const max = (maxRounds - phase1.rounds - phase2.rounds) ? phase1.rounds : 0;
         return max;
     }
 
+    // handles submit and bubbles up phase1-3 data
     const handleSubmit = (e) => {
         e.preventDefault();
         try {
@@ -47,6 +57,7 @@ const Behavior = ({ rounds, onSubmit, onCancel, isClicked }) => {
         debugger;
     }
 
+    // handle's cancel, no logic attached yet
     const handleCancel = (e) => {
         e.preventDefault();
         console.log('cancel called on behavior view.\nCall onCancel if passed.\nNav to previous view if not passed');
@@ -66,11 +77,17 @@ const Behavior = ({ rounds, onSubmit, onCancel, isClicked }) => {
         console.log('phase update');
 
     }
-
+    // should i stop being lazy and separate these into 3 useEffects?
     useEffect(() => {
         handlePhaseUpdates();
     }, [phase1, phase2, phase3])
 
+
+    //TODO
+    /*
+        implment update to behaviorinputcontainer logic
+        refractor form children to 3 behaviorinputcontainer components + the last stuff
+    */
     return (
         <div className="flex flex-row flex-wrap justify-center items-center rounded-md text-slate-100 text-left w-full">
