+++ client/src/components/BehaviorInput.jsx
@@ -1,0 +1,50 @@
+import React, { useState, useEffect } from 'react';
+//import NumberRange from './NumberRange';
+import NumberSelection from './NumberSelection';
+
+//component to organize inputs for a behavior component
+const BehaviorInput = ({ phase, name, orders, rounds, phaseUpdate, disabled }) => {
+    const [newPhase, setNewPhase] = useState(phase);
+
+    const handleRoundsChange = (e) => {
+        const newVal = parseInt(e);
+        setNewPhase({ ...newPhase, rounds: typeof newVal === 'number' ? newVal : parseInt(e.target?.value) });
+    }
+
+    const handleOrdersChange = (e) => {
+        const newVal = parseInt(e);
+        setNewPhase({ ...newPhase, orders: typeof newVal === 'number' ? newVal : parseInt(e.target?.value) });
+    }
+
+    //bubble up with newPhase
+    useEffect(() => {
+        if (newPhase != phase) {
+            phaseUpdate(newPhase)
+        }
+    }, [newPhase]);
+
+    return (
+        <div className='py-1 bg-slate-100'>
+            <NumberSelection
+                name={name}
+                label={rounds.label}
+                min={rounds.min ? rounds.min : 0}
+                max={rounds.max ? rounds.max : 0}
+                val={phase.rounds}
+                setVal={handleRoundsChange}
+                disabled={disabled}
+            />
+            <NumberSelection
+                name={name}
+                label={orders.label}
+                min={orders.min ? orders.min : 0}
+                max={orders.max ? orders.max : 0}
+                val={phase.orders}
+                setVal={handleOrdersChange}
+                disabled={disabled} />
+        </div>
+    )
+
+}
+
+export default BehaviorInput;
